% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_listing_antiquity.R
\name{calculate_listing_antiquity}
\alias{calculate_listing_antiquity}
\title{Calculate Listing Antiquity}
\usage{
calculate_listing_antiquity(data, column_name)
}
\arguments{
\item{data}{The dataset containing the column of interest.}

\item{column_name}{The name of the column in the dataset to process.}
}
\value{
The input dataset with an additional 'years' column that contains the calculated years.
}
\description{
This function extracts the number of days or years from a column in a dataset,
converting days to years where applicable and handling specific strings like "yesterday" and "today".
}
