% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_modified_variables.R
\name{check_modified_variables}
\alias{check_modified_variables}
\title{Compare datasets and identify modified variables}
\usage{
check_modified_variables(current_data, data, count_modified = FALSE)
}
\arguments{
\item{current_data}{The current dataset to compare.}

\item{data}{A list of datasets to compare with 'current_data'.}

\item{count_modified}{Logical indicating whether to generate a count variable for the modified variables.}
}
\value{
The updated 'current_data' dataset with new variables indicating the modified variables and, if specified, the count of modified variables.
}
\description{
This function compares a dataset 'current_data' with one or more datasets 'data' based on the 'url' variable.
It identifies the variables that have been modified for each 'url' in 'current_data' and adds a new variable
indicating the modified variables. Additionally, it can generate a count variable for the modified variables.
}
\examples{
# Create the current dataset
current_data <- data.frame(url = c('url1', 'url2', 'url3'),
                           var1 = c(1, 2, 3),
                           var2 = c(4, 5, 6),
                           var3 = c(7, 8, 9))
current_data

# Output:
#   url var1 var2 var3
# 1 url1    1    4    7
# 2 url2    2    5    8
# 3 url3    3    6    9

# Create the datasets to compare
data1 <- data.frame(url = c('url1', 'url2'),
                    var1 = c(1, 10),
                    var2 = c(4, 5),
                    var3 = c(7, 8))
data2 <- data.frame(url = c('url1', 'url3'),
                    var1 = c(1, 2),
                    var2 = c(4, 50),
                    var4 = c(70, 8))

# Compare datasets using the check_modified_variables function
result <- check_modified_variables(current_data, list(data1, data2), count_modified = TRUE)
result

# Output:
#   url var1 var2 var3 modified_variables modified_count
# 1 url1    1    4    7            var1,var3              2
# 2 url2    2    5    8                 <NA>              0
# 3 url3    3    6    9                 <NA>              0
}
