% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_price_variations.R
\name{check_price_variations}
\alias{check_price_variations}
\title{Check for price variations for a given URL across datasets using specified base dataset(s)}
\usage{
check_price_variations(current_data, previous_data)
}
\arguments{
\item{current_data}{The base dataset for comparison.}

\item{previous_data}{The dataset(s) list to compare against the actual base dataset.}
}
\value{
A dataset with the 'current_data' data and an additional column 'has_new_price' indicating price variations.
}
\description{
This function checks if there are variations in the 'price' variable for a given 'url'
across two or more datasets, using the specified base dataset(s) for comparison.
}
