% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_duplicates.R
\name{export_duplicates}
\alias{export_duplicates}
\title{Export Duplicate Geographic Coordinates}
\usage{
export_duplicates(data, path = getwd())
}
\arguments{
\item{data}{The dataset containing the geographic coordinates.
It should have at least two columns named 'latitude' and 'longitude'
representing the coordinates, and an 'arba_code' column for grouping.}

\item{path}{The path where the output files will be saved. Default is the current working directory.}
}
\description{
This function takes a dataset containing geographic coordinates and exports the duplicated records
for each group to separate GeoPackage files. Each duplicated group is identified by the 'arba_code' column.
}
