% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gpkg.R
\name{export_gpkg}
\alias{export_gpkg}
\title{export_gpkg}
\usage{
export_gpkg(x)
}
\description{
Generates an 'inmo.gpkg' file in the 'output' directory. If the file exists it does not replace it, it must be manually deleted.
}
