% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multiple_gpkg.R
\name{export_multiple_gpkg}
\alias{export_multiple_gpkg}
\title{export_multiple_gpkg}
\usage{
export_multiple_gpkg(x)
}
\description{
Segments the inmoscrap database and generates multiple .gpkg output files according to the ARBA code obtained with the 'spatial_join_ARBA' function
}
