% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_repeated_digits.R
\name{filter_repeated_digits}
\alias{filter_repeated_digits}
\title{Filter a dataframe based on the count of unique digits in a specified column}
\usage{
filter_repeated_digits(data, column)
}
\arguments{
\item{data}{A dataframe containing the data}

\item{column}{The name of the column in the dataframe to apply the function to}
}
\value{
The filtered dataframe where the count of unique digits in the specified column is greater than 1
}
\description{
This function takes a dataframe and filters the data based on the condition
that the count of unique digits is greater than 1 in the specified column.
}
\examples{
# Example dataframe
df <- data.frame(ID = 1:5, Numbers = c(123, 4567, 89, 1010, 222))

# Apply the function to the "Numbers" column and filter the dataframe
filtered_df <- filter_repeated_digits(df, "Numbers")
print(filtered_df)

}
