% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arba_polygons.R
\name{get_arba_polygons}
\alias{get_arba_polygons}
\title{Download and prepare polygons for the province of Buenos Aires from the ARBA website.}
\usage{
get_arba_polygons(include_crown = FALSE)
}
\arguments{
\item{include_crown}{Logical indicating whether to include the "crown" information in the resulting data frame.
If TRUE, a "corona" column will be included to represent the administrative region.
If FALSE (default), the "corona" column will be omitted.}
}
\value{
A spatial data frame containing polygons for the province of Buenos Aires.
}
\description{
This function downloads polygon data for the province of Buenos Aires from the ARBA (Administración
de Ingresos Públicos de la Provincia de Buenos Aires) website. The data is processed and returned as
a spatial data frame suitable for spatial analysis and visualization.
}
\examples{
arba_polygons <- get_arba_polygons()

}
