% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_ovs.R
\name{get_data_ovs}
\alias{get_data_ovs}
\title{Download and Read Soil Data from the Observatorio de Valores de Suelo (https://observatoriosuelo.gba.gob.ar/)}
\usage{
get_data_ovs()
}
\arguments{
\item{None}{}
}
\value{
An 'sf' object containing soil data.

An 'sf' object containing soil data.
}
\description{
This function downloads a .gpkg (Geopackage) file containing soil data from the Observatory of Soil Values
in the Province of Buenos Aires. It then converts the data into an 'sf' (Simple Features) object for further analysis.
}
\details{
The function relies on the 'googledrive' and 'sf' packages. If these packages are not installed, they need to be
installed before using this function.
}
\examples{
\dontrun{
data <- get_data_ovs()
head(data)
}

\dontrun{
data <- get_data_ovs()
head(data)
}

}
\seealso{
\href{https://cran.r-project.org/package=googledrive}{googledrive}

\href{https://cran.r-project.org/package=sf}{sf}
}
