% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_replace.R
\name{missing_replace}
\alias{missing_replace}
\title{missing_replace}
\usage{
missing_replace(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A modified data frame or tibble with empty strings and 'NaN' values replaced by NA.
}
\description{
The function takes a data frame or tibble as input and replaces empty strings ('') and 'NaN' values with NA.
}
\details{
This function replaces empty strings ('') and 'NaN' values in a data frame with NA.
}
\examples{
# Create a data frame with missing values
df <- data.frame(A = c('', 'value', 'NaN'), B = c('value', '', 'NaN'), stringsAsFactors = FALSE)
df

# Output:
#      A     B
# 1         value
# 2 value
# 3 NaN   NaN

# Apply the missing_replace function
modified_df <- missing_replace(df)
modified_df

# Output:
#      A     B
# 1  <NA>  value
# 2 value  <NA>
# 3  <NA>  <NA>
}
