% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_join_arba.R
\name{spatial_join_arba}
\alias{spatial_join_arba}
\title{spatial_join_arba}
\usage{
spatial_join_arba(x)
}
\description{
Spatial union with the ARBA polygon base and recoding with official nomenclature.
}
