% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_join_arba_sf.R
\name{spatial_join_arba_sf}
\alias{spatial_join_arba_sf}
\title{Spatial join with ARBA polygon base and attribute recoding}
\usage{
spatial_join_arba_sf(data, include_crown = FALSE)
}
\arguments{
\item{data}{A spatial object representing points, lines, or polygons.}

\item{include_crown}{Logical, indicating whether to include the 'corona' column. Default is FALSE.}
}
\value{
A modified spatial object with additional attributes obtained from the spatial
join with the ARBA polygon base. If 'include_crown' is TRUE, it also includes a
'corona' column.
}
\description{
This function performs a spatial join between a spatial object and the base of the ARBA
polygon. It also recodes the resulting attributes using the official nomenclature and,
optionally, adds a 'corona' column for the districts of the RMBA
\itemize{
\item Buenos Aires Metropolitan Region - (Fernández Leonardo, 2011. Instituto del Conurbano.
Universidad Nacional de General Sarmiento).
For the district of La Matanza, the function takes into account the criteria
established by INDEC (INDEC, 2003. https://www.indec.gob.ar/dbindec/folleto_gba.pdf).
}
}
\examples{
# Example usage
data <- data.frame(property_group = c('Casa', 'Ph', 'Departamento', 'Casa', 'Ph', 'Departamento'),
land_surface = c(500, 200, 0, 700, 350, 0),
reconstructed_land_surface = c(0, 0, 150, 0, 0, 200),
total_surface = c(700, 300, 800, 600, 400, 1000),
covered_surface = c(600, 250, 750, 550, 350, 900),
bed_amnt = c(5, 3, 2, 6, 4, 2),
bath_amnt = c(3, 2, 1, 4, 2, 1),
is_new_property = c(FALSE, TRUE, FALSE, TRUE, TRUE, FALSE),
year_built = c(1990, 2020, 1980, 2005, 2015, 1995),
garage_amnt = c(2, 0, 1, 3, 0, 2),
room_amnt = c(8, 4, 3, 7, 5, 4),
is_finished_property = c(TRUE, FALSE, TRUE, TRUE, TRUE, FALSE),
latitude = c(-34.6118, -34.6518, -34.9226, -34.5831, -34.6058, -36.7597),
longitude = c(-58.4173, -58.3595, -57.9500, -58.4009, -58.4403, -59.8597)
)
# Perform spatial join without 'corona' column
result <- spatial_join_arba(data)
# Perform spatial join with 'corona' column
result_with_crown <- spatial_join_arba(data, include_crown = TRUE)

}
\references{
More information about ARBA: https://www.arba.gov.ar/

Fernández Leonardo (2011).Censo 2010. Somos 14.819.137 habitantes en
la Región Metropolitana de Buenos Aires. Instituto del Conurbano.
Universidad Nacional de General Sarmiento. www.urbared.ungs.edu.ar

https://www.indec.gob.ar/dbindec/folleto_gba.pdf'
}
