% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_tails.R
\name{trim_tails}
\alias{trim_tails}
\title{Filter a dataframe by trimming the tails of a specified column}
\usage{
trim_tails(data, column, lower_percentage, upper_percentage)
}
\arguments{
\item{data}{The input dataframe}

\item{column}{The name of the column in the dataframe to be trimmed}

\item{lower_percentage}{The percentage of observations to exclude from the lower end of the distribution}

\item{upper_percentage}{The percentage of observations to exclude from the upper end of the distribution}
}
\value{
The filtered dataframe after trimming the specified column
}
\description{
This function trims the tails of a distribution in a specified column of a dataframe,
based on the percentage to exclude from the lower and upper ends of the distribution.
}
\examples{
# Example dataframe
df <- data.frame(ID = 1:10, Values = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

# Filter the dataframe by trimming the "Values" column, excluding 20\% from the lower end and 10\% from the upper end
filtered_df <- trim_tails(df, "Values", 20, 10)
print(filtered_df)

}
