% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{data_quality}
\alias{data_quality}
\title{Filter data based on property type and quality criteria}
\usage{
data_quality(data)
}
\arguments{
\item{data}{The dataset to be filtered.}
}
\value{
A filtered dataset with an additional column "data_quality" indicating the quality category of each property.
}
\description{
This function filters a dataset based on property type and specific quality criteria for different property groups, including houses, PHs, and apartments.
}
\details{
The function applies the following filtering criteria to determine the quality category of each property:

##########warning!!!! ojo!!!! chequear que una vez obtenidas las categorias de esta variable
ver que a la hora de nalizar y operar copn datos considere las columnas correctas( complementariedad de las variables de superficie). es decir, ver que con que variables operar, por ejemplo de superficie (si es reconstructed o sin reconstructed)

\link{Documentation goes here...}
}
\examples{
data <- read.csv("property_data.csv")
filtered_data <- data_quality(data)

}
