% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_join_arba.R
\name{spatial_join_arba}
\alias{spatial_join_arba}
\title{Spatial join with ARBA polygon base and attribute recoding}
\usage{
spatial_join_arba(data)
}
\arguments{
\item{data}{A spatial object representing points, lines, or polygons.}
}
\value{
A modified spatial object with additional attributes obtained from the spatial join with the ARBA polygon base.
}
\description{
This function performs a spatial join between a spatial object and the ARBA polygon base. It also recodes the resulting attributes using the official nomenclature.
}
\examples{
# Example usage
data <- read_sf("path/to/your/data.shp")
result <- spatial_join_arba(data)

}
\references{
More information about ARBA: https://www.arba.gov.ar/
}
